+++ board-game-statistics/src/utils/constants.ts
@@ -1,5 +1,3 @@
-import type { Permission } from './types';
-
 // Page urls
 export const HOME_PAGE_URL = '/';
 export const GAMES_PAGE_URL = '/games';
@@ -12,19 +10,4 @@ export const allowedPagesWhileLoggedOut = [
 	HOME_PAGE_URL,
 	LOGIN_PAGE_URL,
 	CREATE_ACCOUNT_PAGE_URL,
-];
-
-interface PermissionName {
-	permission: Permission;
-	name: string;
-}
-
-export const PermissionNames: PermissionName[] = [
-	{ permission: 'MANAGE_MEMBERS', name: 'Manage members' },
-	{
-		permission: 'MANAGE_MEMBER_PERMISSIONS',
-		name: 'Manage member permissions',
-	},
-	{ permission: 'MANAGE_BOARD_GAMES', name: 'Manage board games' },
-	{ permission: 'MANAGE_GAMES_PLAYED', name: 'Manage games played' },
 ];
